refactor: 빈 생성자 주입 + 필드 타입 변경
- 역직렬화시 기본 생성자가 없어서 에러 발생 -> 생성자 주입으로 해결
- JSON 파싱 시 Array 타입을 String으로 역직렬화하려고 시도해 에러 발생 -> feedImage : List 타입으로 변경 & Hashtag의 id : Long 타입으로 변경

+++ src/main/java/spotlight/spotlight_ver2/repository/FeedRepository.java
@@ -6,5 +6,5 @@ import spotlight.spotlight_ver2.entity.Feed;
 import java.util.List;
 
 public interface FeedRepository extends JpaRepository<Feed, Long> {
-    List<Feed> findByHashtagsId(Integer hashtagId);
+    List<Feed> findByHashtagsId(Long hashtagId);
 }
